refactor: path alias를 이용하여 경로 간결화
