+++ package.json
@@ -29,6 +29,7 @@
     "eslint": "^8.57.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.6",
+    "jest": "^29.7.0",
     "msw": "^2.3.0",
     "ts-node": "^10.9.2",
     "typescript": "^5.2.2",
