+++ src/App.tsx
@@ -1,4 +1,4 @@
-import ProductListPage from "./pages/ProductListPage";
+import ProductListPage from "@pages/ProductListPage";
 
 function App() {
   return (
