+++ src/apis/clients/cartClient.ts
@@ -1,6 +1,6 @@
-import { CART_API_URL, USERNAME, USER_PASSWORD } from "../../env/envVariables";
 import ApiClient from "./ApiClient";
-import { generateBasicToken } from "../__utils__/generateBasicToken";
+import { CART_API_URL, USERNAME, USER_PASSWORD } from "@env/envVariables";
+import { generateBasicToken } from "@apis/__utils__/generateBasicToken";
 
 export const cartClient = new ApiClient(CART_API_URL, {
   Authorization: generateBasicToken(USERNAME, USER_PASSWORD),
