+++ src/components/common/IntersectionDetector.tsx
@@ -1,4 +1,4 @@
-import useIntersectionObserver from "../../hooks/useIntersectionObserver";
+import useIntersectionObserver from "@hooks/useIntersectionObserver";
 
 interface IntersectionDetectorProps {
   onIntersected: () => void;
