+++ src/main.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./App.tsx";
-import { GlobalLayout, GlobalStyle } from "./styles/GlobalStyle.tsx";
+import App from "@src/App.tsx";
+import { GlobalLayout, GlobalStyle } from "@styles/GlobalStyle.tsx";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
