+++ src/pages/ProductListPage.tsx
@@ -1,6 +1,6 @@
-import ProductList from "../components/ProductListPage/ProductList";
-import { CartItemsProvider } from "../store/cartItemsContext";
-import { ErrorToastProvider } from "../store/errorToastContext";
+import ProductList from "@components/ProductListPage/ProductList";
+import { CartItemsProvider } from "@store/cartItemsContext";
+import { ErrorToastProvider } from "@store/errorToastContext";
 
 const ProductListPage = () => {
   return (
